JAMES-2865 move MockSmtpBehaviors to model package
